Create base monthly rule clasS
